+++ src/app/bungie-api/rate-limit-config.ts
@@ -30,4 +30,11 @@ export default function setupRateLimiter() {
   RateLimiterConfig.addLimiter(
     new RateLimiterQueue(/www\.bungie\.net\/Platform\/Destiny2\/Actions\/Items\/EquipItem/, 1, 100)
   );
+  RateLimiterConfig.addLimiter(
+    new RateLimiterQueue(
+      /www\.bungie\.net\/Platform\/Destiny2\/Actions\/Items\/InsertSocketPlugFree/,
+      2,
+      1100
+    )
+  );
 }