+++ src/app/inventory/advanced-write-actions.ts
@@ -1,9 +1,12 @@
 import { currentAccountSelector } from 'app/accounts/selectors';
+import { HttpStatusError } from 'app/bungie-api/http-client';
+import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { t } from 'app/i18next-t';
 import { d2ManifestSelector } from 'app/manifest/selectors';
 import { ThunkResult } from 'app/store/types';
 import { DimError } from 'app/utils/dim-error';
 import { errorLog } from 'app/utils/log';
+import { Destiny2CoreSettings } from 'bungie-api-ts/core';
 import {
   AwaAuthorizationResult,
   AwaType,
@@ -11,6 +14,7 @@ import {
   DestinyItemChangeResponse,
   DestinySocketArrayType,
   insertSocketPlug,
+  insertSocketPlugFree,
 } from 'bungie-api-ts/destiny2';
 import { get, set } from 'idb-keyval';
 import { DestinyAccount } from '../accounts/destiny-account';
@@ -25,37 +29,128 @@ let awaCache: {
   [key: number]: AwaAuthorizationResult & { used: number };
 };
 
+export function canInsertPlug(
+  socket: DimSocket,
+  plugItemHash: number,
+  destiny2CoreSettings: Destiny2CoreSettings,
+  defs: D2ManifestDefinitions
+) {
+  return $featureFlags.awa || canInsertForFree(socket, plugItemHash, destiny2CoreSettings, defs);
+}
+
+function canInsertForFree(
+  socket: DimSocket,
+  plugItemHash: number,
+  destiny2CoreSettings: Destiny2CoreSettings,
+  defs: D2ManifestDefinitions
+) {
+  const { insertPlugFreeProtectedPlugItemHashes, insertPlugFreeBlockedSocketTypeHashes } =
+    destiny2CoreSettings;
+  if (
+    (insertPlugFreeProtectedPlugItemHashes || []).includes(plugItemHash) ||
+    (insertPlugFreeBlockedSocketTypeHashes || []).includes(socket.socketDefinition.socketTypeHash)
+  ) {
+    return false;
+  }
+
+  const plug = defs.InventoryItem.get(plugItemHash);
+
+  const free =
+    // Must be reusable or randomized type
+    Boolean(
+      socket.socketDefinition.reusablePlugItems.length > 0 ||
+        socket.socketDefinition.reusablePlugSetHash ||
+        socket.socketDefinition.randomizedPlugSetHash
+    ) &&
+    // And have no cost to insert
+    !plug.plug?.insertionMaterialRequirementHash &&
+    // And the current plug didn't cost anything (can't replace a non-free mod with a free one)
+    (!socket.plugged || !socket.plugged.plugDef.plug.insertionMaterialRequirementHash);
+
+  return free;
+}
+
+/**
+ * Modify an item to insert a new plug into one of its socket.
+ */
 export function insertPlug(item: DimItem, socket: DimSocket, plugItemHash: number): ThunkResult {
   return async (dispatch, getState) => {
     const account = currentAccountSelector(getState())!;
 
+    const free = canInsertForFree(
+      socket,
+      plugItemHash,
+      getState().manifest.destiny2CoreSettings!,
+      d2ManifestSelector(getState())!
+    );
+
     // The API requires either the ID of the character that owns the item, or
     // the current character ID if the item is in the vault.
     const storeId = item.owner === 'vault' ? currentStoreSelector(getState())!.id : item.owner;
-    try {
-      const actionToken = await getAwaToken(account, AwaType.InsertPlugs, storeId, item);
-      // TODO: if the plug costs resources to insert, add a confirmation. This'd
-      // be a great place for a dialog component?
 
-      const response = await insertSocketPlug(authenticatedHttpClient, {
-        actionToken,
-        itemInstanceId: item.id,
-        plug: {
-          socketIndex: socket.socketIndex,
-          socketArrayType: DestinySocketArrayType.Default,
-          plugItemHash,
-        },
-        characterId: storeId,
-        membershipType: account.originalPlatformType,
-      });
+    try {
+      const insertFn = free ? awaInsertSocketPlugFree : awaInsertSocketPlug;
+      const response = await insertFn(account, storeId, item, socket, plugItemHash);
 
       // Update items that changed
       await dispatch(refreshItemAfterAWA(item, response.Response));
     } catch (e) {
       errorLog('AWA', "Couldn't insert plug", item, e);
-      showNotification({ type: 'error', title: t('AWA.Error'), body: e.message });
+      if (e instanceof DimError && e.cause instanceof HttpStatusError && e.cause.status === 404) {
+        showNotification({
+          type: 'error',
+          title: 'Not Yet!',
+          body: "Changing perks and mods won't be available until the launch of the Bungie 30th Anniversary patch on December 7th.",
+        });
+      } else {
+        showNotification({ type: 'error', title: t('AWA.Error'), body: e.message });
+      }
     }
   };
+}
+
+async function awaInsertSocketPlugFree(
+  account: DestinyAccount,
+  storeId: string,
+  item: DimItem,
+  socket: DimSocket,
+  plugItemHash: number
+) {
+  return insertSocketPlugFree(authenticatedHttpClient, {
+    itemId: item.id,
+    plug: {
+      socketIndex: socket.socketIndex,
+      socketArrayType: DestinySocketArrayType.Default,
+      plugItemHash,
+    },
+    characterId: storeId,
+    membershipType: account.originalPlatformType,
+  });
+}
+
+async function awaInsertSocketPlug(
+  account: DestinyAccount,
+  storeId: string,
+  item: DimItem,
+  socket: DimSocket,
+  plugItemHash: number
+) {
+  const actionToken = await getAwaToken(account, AwaType.InsertPlugs, storeId, item);
+
+  // TODO: if the plug costs resources to insert, add a confirmation. This'd
+  // be a great place for a dialog component?
+
+  return insertSocketPlug(authenticatedHttpClient, {
+    actionToken,
+    itemInstanceId: item.id,
+    plug: {
+      socketIndex: socket.socketIndex,
+      socketArrayType: DestinySocketArrayType.Default,
+      plugItemHash,
+    },
+    characterId: storeId,
+    membershipType: account.originalPlatformType,
+  });
 }
 
 /**