+++ src/browsercheck-utils.js
@@ -24,7 +24,7 @@ export const unsupported = {
   it: "Il team di DIM non supporta l'utilizzo di questo browser. Alcune o tutte le funzioni di DIM potrebbero non funzionare.",
   ja: 'DIM チームはこのブラウザの利用をサポートしていません。DIM の一部または全ての機能が動作しない可能性があります。',
   ko: 'DIM 팀은 이 브라우저를 지원하지 않습니다. DIM의 일부나 전체 기능이 동작하지 않을 수 있습니다.',
-  pl: 'The DIM team does not support using this browser. Some or all DIM features may not work.',
+  pl: 'Zespół DIM nie obsługuje tej przeglądarki. Niektóre lub wszystkie funkcje DIM mogą nie działać.',
   'pt-br':
     'A equipe do DIM não oferece suporte a este navegador. Alguns recursos do DIM podem não funcionar.',
   ru: 'The DIM team does not support using this browser. Some or all DIM features may not work.',
@@ -43,7 +43,7 @@ export const steamBrowser = {
   it: 'Sembra che questa pagina sia caricata tramite il browser di Steam. A causa delle sue risorse e caratteristiche limitate, potrebbe non riuscire ad eseguire DIM correttamente. Non possiamo fornire supporto a riguardo.',
   ja: 'Steam ブラウザにこのページが読み込まれているようです。 機能やリソースが限られているため、予期しないまたは断続的に DIM の実行に失敗する可能性があります。サポートを提供できません。',
   ko: '이 페이지가 Steam 브라우저에서 로드된 것 같습니다. 한정된 기능과 자원으로 인해, 예기치 않게 혹은 간헐적으로 DIM을 실행에 실패할 수 있습니다. 이에 대한 지원은 제공되지 않습니다.',
-  pl: "It looks like this page is loaded in Steam's browser. Due to its limited features and resources, it may unexpectedly or intermittently fail to run DIM. We cannot provide support for it.",
+  pl: 'Wygląda na to, że ta strona jest załadowana w przeglądarce Steam. Ze względu na ograniczone funkcje i zasoby, DIM może nie działać. Nie możemy zapewnić wsparcia dla tej przeglądarki.',
   'pt-br':
     'Parece que esta página foi aberta no navegador do Steam. Devido à limitações, a execução do DIM pode falhar de forma inesperada ou intermitente. Nós não podemos oferecer suporte a ele.',
   ru: "It looks like this page is loaded in Steam's browser. Due to its limited features and resources, it may unexpectedly or intermittently fail to run DIM. We cannot provide support for it.",