+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Loadouts can now have notes.
 * Share loadout build settings (mods, notes, loadout optimizer settings) from the Loadouts page.
+* Loadouts can now save stasis subclass abilities, aspects, and fragments.
 * The menu for pinning or excluding an item in Loadout Optimizer now only shows items that match the overall search filter.
 * Stat searches support keywords like "highest" and "secondhighest" in stat total/mean expressions. e.g. basestat:highest&secondhighest:>=17.5